+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_09_03_125812) do
+ActiveRecord::Schema[8.1].define(version: 2026_09_17_120000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -255,6 +255,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_09_03_125812) do
     t.datetime "created_at", null: false
     t.string "hint_cy"
     t.string "hint_en"
+    t.string "language"
     t.string "options_cy", default: [], array: true
     t.string "options_en", default: [], array: true
     t.integer "position", null: false